Migration 64 bits

+++ ColorPicker/NPColorPickerView.m
@@ -465,7 +465,7 @@ NSString * kColorProperty = @"color";
 
    CGFloat teta = M_PI_3 * saturation;
 
-   CGFloat l = brightness * ( (sqrtf(3) / 2) / cosf(fabsf(teta - (M_PI/6))));
+   CGFloat l = brightness * ( (sqrtf(3) / 2) / cosf(fabs(teta - (M_PI/6))));
    CGPoint s = (CGPoint) { center.x + cosf(2*M_PI_3) * internalRadius, center.y + sinf(2*M_PI_3) * internalRadius };   
    CGPoint b = (CGPoint) { center.x + cosf(4*M_PI_3) * internalRadius, center.y + sinf(4*M_PI_3) * internalRadius };
    
@@ -518,7 +518,7 @@ NSString * kColorProperty = @"color";
    
    if (brightness) {
       CGFloat v1l = sqrtf(v1.x*v1.x + v1.y*v1.y);
-      CGFloat max = v1l  * ( (sqrtf(3) / 2) / cosf(fabsf(teta - (M_PI/6))));
+      CGFloat max = v1l  * ( (sqrtf(3) / 2) / cosf(fabs(teta - (M_PI/6))));
       CGFloat l = sqrtf(v2.x*v2.x + MIN(v2.y,0)*MIN(v2.y,0));
       *brightness = MIN( l, max) / max ;
    }
@@ -551,7 +551,7 @@ NSString * kColorProperty = @"color";
       CGPoint center = CGPointMake(CGRectGetMidX(viewFrame), CGRectGetMidY(viewFrame));
       CGPoint t = [recognizer locationOfTouch:0 inView:self];
       
-      float sat, brigt, hue = (M_PI - atan2f(t.y-center.y,center.x-t.x)) / (2 * M_PI);
+      CGFloat sat, brigt, hue = (M_PI - atan2f(t.y-center.y,center.x-t.x)) / (2 * M_PI);
       [color_ getHue:NULL saturation:&sat brightness:&brigt alpha:NULL];
       [self setColor:[UIColor colorWithHue:hue saturation:sat brightness:brigt alpha:1.0f]];
    } else if ([recognizer state] == UIGestureRecognizerStateEnded) {
@@ -575,7 +575,7 @@ NSString * kColorProperty = @"color";
 -(void)onMoveSVIndicator:(UIPanGestureRecognizer *) recognizer {
    if ([recognizer state] == UIGestureRecognizerStateBegan || [recognizer state] == UIGestureRecognizerStateChanged) {
       CGPoint t = [recognizer locationOfTouch:0 inView:self];
-      float sat, brigt, hue;
+      CGFloat sat, brigt, hue;
       [color_ getHue:&hue saturation:NULL brightness:NULL alpha:NULL];
       [self getSaturation:&sat brightness:&brigt position:CGPointMake(t.x,t.y)];
       [self setColor:[UIColor colorWithHue:hue saturation:sat brightness:brigt alpha:1.0f]];
